refactor(broker): split order_pricer.rs into directory layout (C30 of PR-3)
Pure mechanical move. order_pricer.rs (3047 lines) becomes a directory:

- order_pricer/mod.rs (~41 lines): module decls + pub use service::OrderPricer
  + #[cfg(test)] pub(crate) use service::tests so external test consumers
  (shared/prioritization.rs) keep working at crate::order_pricer::tests::*.
- order_pricer/error.rs (~48 lines): broker-specific skip-code constants
  (SKIP_NOT_IN_ALLOWLIST through SKIP_PRICING_ERROR), the OrderPricerErr alias,
  and the manual impl CodedError for OrderPricingError (foreign #[non_exhaustive]
  type, so coded_error_impl! doesn't fit).
- order_pricer/types.rs (~129 lines): ActivePreflights cancellation tracker
  with PreflightHandle, OrderCache type alias, cache size/TTL constants,
  and the format_truncated helper used by ActivePreflights::format.
- order_pricer/service.rs (~2501 lines): OrderPricer<P> struct (fields
  promoted to pub(super) for cross-module access from helpers.rs), pub fn new,
  impl BrokerService for OrderPricer<P>, and the original test module.
- order_pricer/helpers.rs (~470 lines): impl<P> OrderPricer<P> with
  price_order_and_update_state and gas_balance_reserved, plus the entire
  impl<P> OrderPricingContext for OrderPricer<P> (the trait hook surface
  that backs upstream price_order).
- order_pricer/price_oracle.rs (~36 lines): absorbed from top-level
  crates/broker/src/price_oracle.rs without modification.

External path crate::order_pricer::OrderPricer::new preserved via re-export.
crate::order_pricer::tests::{OrderParams, PricerTestCtxBuilder} preserved
via the cfg-test re-export. lib.rs has its mod price_oracle; line removed.
Test module gains explicit imports it used to inherit via super::*:
ActivePreflights, MarketConfig, now_timestamp, OrderPricingContext,
OrderPricingOutcome.
